+++ examples/basic-vite-ssr/src/app.tsx
@@ -1,0 +1,13 @@
+import { runApp, IAppConfig } from 'ice';
+
+const appConfig: IAppConfig = {
+  app: {
+    rootId: 'ice-container',
+    errorBoundary: true,
+  },
+  router: {
+    type: 'browser',
+  },
+};
+
+runApp(appConfig);