+++ examples/basic-vite-ssr/src/pages/Dashboard/Page2/index.tsx
@@ -1,0 +1,5 @@
+export default () => {
+  return (
+    <div>Page2</div>
+  );
+};