+++ packages/plugin-ice-ssr/src/env.ts
@@ -1,0 +1,2 @@
+// setup server env
+(global as any).__IS_SERVER__ = true;